+++ database/migrations/2020_01_01_000005_create_activity_log_table.php
@@ -25,7 +25,7 @@ class CreateActivityLogTable extends Migration
             $table->timestamps();
 
             // Indexes
-            $table->index('log_name');
+            $table->index(['log_name']);
         });
     }