+++ resources/views/adminarea/pages/datatable-dropzone.blade.php
@@ -26,7 +26,7 @@
                         {{ Form::open(['url' => $url, 'class' => 'dropzone', 'id' => "{$id}-media-dropzone", 'data-dz-accepted-files' => 'application/vnd.ms-excel']) }}
                             <div class="dz-message" data-dz-message><span>{{ trans('cortex/foundation::common.drop_to_import') }}</span></div>
                         {{ Form::close() }}
-                        {!! $dataTable->table(['class' => 'table table-striped table-hover responsive dataTableBuilder', 'id' => "{$id}"]) !!}
+                        {!! $dataTable->pusher($pusher ?? null)->table(['id' => $id]) !!}
                     </div>
 
                 </div>