+++ resources/views/managerarea/pages/datatable-logs.blade.php
@@ -23,7 +23,7 @@
                 <div class="tab-content">
 
                     <div class="tab-pane active" id="logs-tab">
-                        {!! $dataTable->table(['class' => 'table table-striped table-hover responsive dataTableBuilder', 'id' => "{$id}"]) !!}
+                        {!! $dataTable->pusher($pusher ?? null)->table(['id' => $id]) !!}
                     </div>
 
                 </div>